// frontend/src/api/openapi/models/CodeNodeData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CodeNodeData = {
  /**
   * Background color of the text
   */
  bgcolor: string;
  /**
   * Background color alpha of the text
   */
  bgalpha: number | null;
  type: string;
  /**
   * ID of the code
   */
  codeId: number;
  /**
   * ID of the parent code
   */
  parentCodeId?: number | null;
};