// frontend/src/api/openapi/models/MemoNodeData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type MemoNodeData = {
  /**
   * Background color of the text
   */
  bgcolor: string;
  /**
   * Background color alpha of the text
   */
  bgalpha: number | null;
  type: string;
  /**
   * ID of the memo
   */
  memoId: number;
};