// frontend/src/api/openapi/models/SpanAnnotationNodeData.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SpanAnnotationNodeData = {
  /**
   * Background color of the text
   */
  bgcolor: string;
  /**
   * Background color alpha of the text
   */
  bgalpha: number | null;
  type: string;
  /**
   * ID of the span annotation
   */
  spanAnnotationId: number;
};